requirements.txt
python-dotenv~=1.0.1

Django~=5.0.3
Pillow

django-allauth~=0.61.1
django-debug-toolbar
django-environ
django-extensions
djangorestframework
django-filter
django-ckeditor-5
django-mptt
pytils~=0.4.1

celery~=5.3.6
redis
